requirements.txt
discord
pylast
python-dotenv